sync pipe: Close pipe before waiting for child, better error message
On an error in getting the stats, close the pipes before waiting on
the child (because some errors don't make the child close, especially
on stats where the child runs continuously.)

Improve the error message - we can't try to show the message including
the header as a string, because the first byte of the 3 byte length
will likely be a '\0' and terminate the string, so the information
isn't useful. Explain that the problem is that the buffer is too
small, and output the partial message. Yes, eventually we should
handle this with multiple reads.