Qt: Parent the QColorDialog in ColoringRulesDialog
On Linux (at least Wayland + GTK3), a Qt applicationModal dialog is
not necessarily always-on-top on the application. If parented, it
is always on top of the parent.

Parent the QColorDialog, which at least makes it stay on top of
the ColoringRulesDialog. ColoringRulesDialog itself, like other
ApplicationModal GeometryStateDialogs, also ought to stay on top
of the application, but we intentionally don't parent GSDs on
Linux (or Windows) because that makes them always-on-top even
when not modal.

Part of #19099. The issue for the modal GeometryStateDialogs
in general, including the ColoringRulesDialog, needs more work.