IEC101/104: Add some expert infos
For IEC-60870-5-101, serial to TCP converters can introduce errors.
Flag the errors better.

Add some expert infos for bogus packets.

When the duplicate length or type fields don't agree, just
take the rest of the current packet, show errors, and try
to sync on a PDU later. We can't really reject a packet
once we're inside tcp_dissect_pdus, and we can't decide which
of the two values for length or type is correct (if they used
three we could take a majority vote).

Fix #19921.