Pass a set of PIDL flags, not "is open" and "is close" Booleans.
Pass to various routines policy handle routines a set of PIDL flags that
indicate, among other things, whether this is opening a handle, closing
a handle, or just using a handle, rather than passing two Booleans for
"it's an open" and "it's a close".

This makes the calls a bit easier to read, as they indicate what the
parameter means.

Add a PIDL_POLHND_USE value of 0, to use in cases where it's neither
opening nor closing a handle, to make calls that do so clearer.