USBLL: Reassemble High-Bandwidth interrupt endpoint data
Assume transfer length up to total payload length on High-Bandwidth
interrupt endpoints if transfer length is not known.

New transfer reassembly starts on a new (not retransmitted) packet when
there is no active reassembly in progress or when the previous packet
did end the transfer. Correct the check for previous packet transfer end
for cases where the transfer length is Max Packet Size multiple.