Media Type: Add a structured suffix table
Add the registered structured syntax suffixes listed at
https://www.iana.org/assignments/media-type-structured-suffix/media-type-structured-suffix.xhtml
for which we have dissectors to a media type suffix table.

Call this from the fallback "nothing else matched" media protocol,
which is used in some dissectors (like HTTP), though not everywhere
that uses the media type table (e.g., not HTTP/2).

Perhaps all dissectors that dissect media types should call the
Media pseudo-protocol, similar to the Ethertype pseudo-protocol.
This would have the advantage of making Decode As work better,
as with Ethertype. (Right now there is a Decode As for media types
in the OBEX dissector, which affects the global table but is only
populated in the Decode As dialog if OBEX is present.)

Fix #19945