pidl: fix the check for Windows.
There's no need to put the string for the OS into a variable; just test
it in an if statement.

Wrap the if condition in parentheses; it appears to be necessary, at
least as I read https://perldoc.perl.org/perlsyn#Compound-Statements
(and if it's not there, I get a syntax error).

According to https://perldoc.perl.org/variables/$%5EO, $^O has the value
"MSWin32" on Windows, not "win".

"eq", not "=" or "==", is the string comparison operator.

Make tools/pidl/idl.yp's parse_file() match the one in
tools/pidl/lib/Parse/Pidl/IDL.pm.