monero: Stop dissection on an invalid type
The length of a value is determined by its type. If the type
is unknown, stop dissecting because we can't know where the
next type begins, and errors will just compound. In particular,
for arrays (which are all one type), we need to abort before
looping, rather than add an expert info for each supposed array
member (especially since the invalid types don't try to fetch
data from the tvb and thus wouldn't throw an exception, instead
looping up to the maximum varint size.)

Fix #19961