wslua: Add some clarification to dofile
Call out that dofile searches the personal configuration directory
and the global configuration directory, which are not the plugin
directories, because that has been a point of confusion.
These directories are *not* added to the package.path search path for
require(), but the plugin directories *are* added to the search path for
require(), but not to the path searched by dofile.

This is largely because the wslua_dofile implementation predates
automatically loading all .lua files in the plugin directories,
updating the search path for require(), and loading init.lua from
the plugin directories instead of the configuration directory.
Those changes each came several years apart.

Move dofile before loadfile, as it comes earlier alphabetically
and the loadfile documentation refers to dofile.

Ping #18589