packetBB: Skip over the extended type before retrieving length
The packetBB TLV parsing parses the header twice, once to get
the length in order to add the tree item with the correct length
(rather than calling proto_item_set_len or similar later), and
once actually adding the items. In the initial retrieval of the
length it fails to increment the offset if the extended type is
present according to the flags bit.

Fix #19972