wslua: Use require to load Lua plugins in plugins directories
Load Lua plugins in the global and personal plugins directories
with require, which adds them to the package.loaded path so they
are only run once (whether loaded by us or by other script earlier
in ASCIIbetical order require()ing it first).

When loading scripts with require() (or dofile()), the recommended
Lua style is to return a function table and have the requiring
script use that. Globals are possible, but deprecated.

Don't have separate user environments for personal plugins, because
we don't do that for global plugins, and it confuses users. This
does mean that personal plugins can pollute the global environment
and there are problems if they use the same global variable. Avoiding
that will be up to the user; this was already the case for plugins
in the global plugins directory.

For command line scripts, continue to put them in a file environment.
Those are handled separately because they support command line options.

Document, including examples.

Fix #18589