rpm-setup: Add correct SUSE Qt6 package names
SUSE has somewhat different package names for Qt6 modules than
RH/Fedora.

We still don't build with Qt6 on SUSE by default. Note that if we
do so, we have to install and specify a more recent version of gcc,
because the default gcc on SUSE 15.5 and 15.6 is gcc-7.5.0, which
will fail compiling against Qt6 due to C++17 limitations.
( <filesystem> is only supported as <experimental/filesystem>
https://gcc.gnu.org/onlinedocs/libstdc++/manual/status.html )