USB: Store endpoint with direction flag in usb_conv_info
Commit 9605f3d3f44f ("USB: Differentiate IN and OUT endpoints") solved
the issue where IN and OUT endpoints with same numbers were treated as
the same endpoint. However, the commit failed to make sure that the
endpoint stored in usb_conv_info includes direction. This lead to
dissectors that do store endpoint numbers during descriptor processing
to no longer be able to match data on IN endpoints. In USB MSC UASP case
this lead to Data In and Status no longer being dissected.

Fix the issue by storing endpoint together with direction flag.