Qt: Don't parent a Uat preference dialog with a menu that gets destroyed
A ProtocolPrefenceMenu containing a UatPreferenceAction opens a
modal UatDialog and then the Menu is deleted.

On some Linux Wayland verisons at least, parenting the UatDialog for a
ProtocolPreferenceMenu with a QMenu that gets deleted causes the
dialog to be deleted and never shown. (There might be some ordering
issues.)

It's an ApplicationModal dialog, so just parent it with the mainWindow.

This prevents UAT preferences from not being able to be selected by
right-clicking on the packet list or proto tree.

Related to 67daaa34213bb10d493fcf88dfc21587e02073c3
(which exposed the issue by parenting modal dialogs on Linux, which
is necessary on Wayland to make them properly on top)