epan: optimize module_check_valid_name
This is called over 250,000x on startup. Benchmarks and call analysis
confirm it is faster on all tested platforms to declare static arrays
with 256 members and only index into it with each byte, than to test
if the byte is < 128 (i.e., ASCII) and index into a smaller array.
This is the way the various glib g_ascii_is functions test for the
same reason.