Add option to save files in current working directory
Note that this is somewhat less than useful when run from
the Finder on macOS (returns the root directory) or from
a shortcut (including the taskbar) on Windows (returns whatever
is set as the fixed "run in" directory for the shortcut, which
is usually the directory where the program resides, certainly
for the shortcut the installer sets up.)

We could try to do the same kind of workaround as done in
get_persdatafile_dir() and for the other file open style settings.
OTOH, this isn't the default setting (as platform UI guidelines
recommend against this sort of thing on macOS and Windows), so
perhaps caveat emptor is ok here.

Fix #1237