GIOP: Don't default to looking for IOR.txt in the CWD
It's unpredictable (and potentially a security issue in the
case of malicious files) to have the dissector look for a file
of a given name in the current working directory by default.
Like the other filename preferences, make the user specify it
in order to use its contents to supplement the dissection.

Fix #15384