Qt: Add GUI support for saving in LZ4 format
Add GUI support for saving in LZ4 format as well as gzip and
uncompressed. Replace the current checkbox with a group box.
The group box looks better alongside the packet range group box
for Export Packet Dissections and would be appropriate to
substitute into the Capture Options dialog in a later commit;
a combobox might look more natural for the ordinary Save As window.

Change the work to fix up the file extension a bit, so that
it can switch between .gz and .lz4