asn2wrs: Create types for type refs with extra value constraints in PER
In PER (and OER) additional value constraints can change the size on
the element in a similar way to size constraints. Add constrained
types when value constraints are added.

If a subtype (Type_Ref) has a value constraint and is an integer
type in PER or OER, use the min and max values from the ValueRange
of the subtype instead of the original type.

Remove some conformance lines that are no longer needed.

Add a conformance line for one still unhandled case, where the
ValueRange constraint is given using names from the NamedNumberList
of the base type.

Fix #19886


(cherry picked from commit a21bada17650ea2e5d7f51fecfdeaec3db1bfe51)

Co-authored-by: John Thacker <johnthacker@gmail.com>