extcap: Ensure command line options are applied
We have to read the preference files and the command line options
before loading the extcaps so that we can see the state of the
capture.no_extcap preference and avoid the (possibly time-consuming)
loading of the extcaps if the user has asked us not to do so.

However, after loading the extcaps, we have to make sure that any
extcap options given on the command line are actually applied.

Fix #20054


(cherry picked from commit b68a80ef4d4f01210c51c24d2a7efa1f8e030046)

Co-authored-by: John Thacker <johnthacker@gmail.com>