tshark: Change fields glossary report with prefix syntax
Backport part of the glossary report changes from the master branch.
Adding the prefix as an extra option separated by a space made
other options, like the configuration profile setting, not work.
Change the syntax to "-G fields,<prefix>", similar to the "-z"
options and as done in the master branch.

Among other things, this makes it easier for people to list the
fields corresponding to columns. (See #20096.)