Don't try to use O_PATH when it's not available.
The O_PATH flag is used in the DBus-based "Open in Folder" functionality on
non-Windows non-Mac platforms.

There exist non-Windows non-Mac platforms which support DBus and lack O_PATH,
and the build would fail on those platforms.

Fall back to using O_RDONLY instead of O_PATH when O_PATH is not available.
This will require that read permission be granted for the file in question, but
that should usually be the case.  As it is, this usage is only on a fallback
path which should only be used when the application is sandboxed.

Fixes #20031


(cherry picked from commit 753f0077edf17fd7abe6f4f28995cd08920295c1)

Co-authored-by: Darius Davis <darius-wireshark@free-range.com.au>