Qt: Allow saving files with Qt 6.8
Prior to Qt 6.8, QFileDialog::selectedNameFilter() returns the current
text displayed as the filter in the dialog; i.e., as we set
HideNameFilterDetails it does not include the parenthetical extension
list.

As of 6.8, it includes the hidden details:
https://bugreports.qt.io/browse/QTBUG-127924

This causes hash lookups for the file type to fail, reporting
unknown file type.
One simple approach is just to add both to the hash.


(cherry picked from commit 702116d425a1542c0db84e17236dae8c7e1b0bfd)

Co-authored-by: John Thacker <johnthacker@gmail.com>