Qt: Don't disconnect all signals (inc. ::destroyed) of WiresharkDialog
As the QObject documentation warns:
"Disconnecting all signal-slot connections will also disconnect the
QObject::destroyed() signal if it is connected. Doing so can adversely
affect classes that rely on this signal for cleaning up resources. It
is recommended to disconnect only the specific signals that were
connected by application code."

In WiresharkDialog we also change the WindowTitle. In Qt 6.8.0,
changing the WindowTitle also communicates the new title to the
AccessibilityWidget.
(https://codereview.qt-project.org/c/qt/qtbase/+/553568)

The AccessibilityWidget is removed from an internal Qt cache when
the widget is destroyed, which doesn't happen if the signals are
disconnected. If a subsequent WiresharkDialog is created with the
same name as a destroyed one (e.g., two Follow Stream Dialogs),
this can result in a lookup in the cache returning the
AccessibilityWidget pointing to the destroyed QDialog (now a
nullptr, thanks to shared pointer), which gets dereferenced and
crashes.

Don't disconnect all signals, which includes internal Qt signals
that happen when the dialog is destroyed. Only disconnect the signal
we added. It's unclear if we need to disconnect signals added by
derived classes (or this one, for that matter), but derived classes
should do that if so.

Thanks to Michael Weghorn for the analysis.

Fix #20174


(cherry picked from commit 7a7b99e0d96ff378f531b40300ce41b65a2c049a)

Co-authored-by: John Thacker <johnthacker@gmail.com>