capture_opts: Only read the list of capture filters when necessary
Instead of reading the list of capture filters from the file whenever
a -f option is given, only do so if the "predef" option is given
indicating that a named capture filter from the list is needed (and
they haven't been read yet.)

Never read the list of named capture filters in capture child mode,
because that processing should have already been done in the calling
program, and the capture filter name should have been replaced with
the real filter already.

For now we still attempt to read the list of named capture filters when
running dumpcap standalone from the command line, but only if -f is
called with "prefdef:<name>" as the option - since dumpcap doesn't
understand profiles, this will always happen from the default profile
location. That shouldn't happen either, but we need a better way to
determine if we're running from dumpcap.

Ping #20108 - this fixes the vast majority of cases where users would
encounter that issue. Separately, dumpcap needs to register libwsutil
report message routines in case some other warning gets reported, but
as it can't register the wiretap-related ones, that's harder to fix
without sacrificing ABI and API compatibility, and this needs to get
fixed in 4.4


(cherry picked from commit 664a9495bfcbb4150a2551676d5075d0acaa6c73)

Co-authored-by: John Thacker <johnthacker@gmail.com>