Lua: Support comma separated DissectorTable ranges in decimal comma locales
For numeric literals, Lua only uses "." as the decimal separator (as,
e.g., the packet matching grammar for display filters does), but
when converting strings back and forth, the locale native separator
is used, unless Lua is specifically changed when compiled. On Lua 5.3
and later, "." can always be used when converting from a string as well.

The functions in epan/range.c use "," to separate ranges of integers,
which are converted back and forth from strings.

On locales which use "," as a decimal separator, this creates ambiguity;
strings like "10,11" can either be converted to the numeric literal of
ten and eleven-hundredths, or a range with two subranges, 10 and 11.

To avoid ambiguity, only convert strings to ranges; do not attempt to
convert strings to numbers/integers first. This will work fine for
strings consisting of a single integer like "10", but does mean than
on English and other locales that use dot as the decimal separator,
"10.0" will no longer be allowed (the numeric literal 10.0 will still
work, because the conversion is exact, see below.)

On Lua 5.3 and later, conversions to an integer will return a failure
if the conversion is not exact (again, unless overridden by a
compilation option when compiling Lua, something we can't control when
UN*X system Lua packages are used). On earlier Lua, floating point
Numbers are simply cast to integer, truncating. This causes silent
unexpected behavior on Wireshark 4.2 and earlier.

Note we don't support (and have never done so) passing in our user
UInt64 type (or the less relevant now Int64 type) to these functions,
though if 64 bit dissector tables are ever implemented (see #20207)
that may be a consideration.

Fix #20216


(cherry picked from commit fb64334c6a6a2aff0a04c6d45c26e9349af71a57)

Co-authored-by: John Thacker <johnthacker@gmail.com>